+++ NEWS.md
@@ -63,6 +63,10 @@
 * Filter ordering is always specified by the keyword `order` and is `FORWARD_FILTER` by
   default.
 
+* In most out-of-place methods, keyword `eltype` may be used to specify the type of the
+  elements of the result.
+
+
 ## Version 2.0.0
 
 Version 2 of `LocalFilters` better integrates in the Julia ecosystem as fewer custom types